Moved the geopolitical units to their own location.  This changed the calls in the `query.js` file and cleaned up the indexing as well.  Added the required sql files.
Other cleanup in the `index` file includes cleaning out unused functions in the `swagger` code blocks.

Cleanup in `queries` moved the new alls to their own folders to help manage sql files.
