+++ helpers/geopoliticalunits/gpubyid.sql
@@ -1,0 +1,2 @@
+SELECT * FROM ndb.geopoliticalunits AS gpu
+WHERE gpu.geopoliticalid = ($1:csv)
