+++ routes/index.js
@@ -19,117 +19,6 @@ router.get('/', function(req, res, next) {
   //res.render('index', { title: 'Express' });
 });
 
-/* (Approximate) LINE NUMBERS & Status:
-[ Use CTRL-G + Line number in Sublime Text ]
-
-  36   - chronology
-  123  - contacts
-  187  - datasets
-  234  - dbtablessites
-  267  - downloads
-         DownloadsCSV
-         DownloadMultiple
-         DownloadMultipleCSV
-         DownloadZip
-  301  - GeoPoliticalUnits
-  339  - occurrence
-  419  - pollen
-  529  - publications
-         SampleData
-  586  - site
-         sites
-  585 - taxa
-
-*/
-
-
-/**
-
-* @swagger
-* definitions:
-*   chronology:
-*     properties:
-*       ChronologyID:
-*         type: integer
-*         format: int64
-*       controls:
-*         type: object
-*         properties:
-*           ChronControlID:
-*             type: integer
-*             format: int64
-*           Age:
-*             type: number
-*             format: float
-*           AgeYounger:
-*             type: number
-*             format: float
-*           AgeOlder:
-*             type: number
-*             format: float
-*           ControlType:
-*             type: string
-*           Depth:
-*             type: number
-*             format: float
-*           Thickness:
-*             type: number
-*             format: float
-*       Default:
-*         type: boolean
-*       ChronologyName:
-*         type: string
-*       AgeType:
-*         type: string
-*       AgeModel:
-*         type: integer
-*         format: int64
-*       AgeYounger:
-*         type: number
-*         format: float
-*       AgeOlder:
-*         type: number
-*         format: float
-*       Notes:
-*         type: string
-*       DatePrepared:
-*         type: string
-*         format: date
-*       Datasets:
-*         type: object
-*         properties:
-*           DatasetID:
-*             type: integer
-*             format: int64
-*           DatasetType:
-*             type: string 
-*/
-
-/**
-* @swagger
-* /v2/data/chronology:
-*   get:
-*     summary: Chronology metadata for a dataset.
-*     description: Returns the chronology and chronological controls used for a dataset age model.
-*     parameters:
-*       - name: chronid
-*         description: Unique chronology identifier.
-*         in: path
-*         required: true
-*         type: string
-*     produces:
-*       - application/json
-*     responses:
-*       200:
-*        description: Chronology
-*        schema:
-*          type: object
-*          items:
-*            $ref: '#/definitions/chronology'
-*/
-router.get('/v2/data/chronology', db.chronology);
-router.get('/v2/data/chronology/:id', db.chronology);
-
 /**
 * @swagger
 * definitions:
@@ -198,9 +87,7 @@ router.get('/v2/data/contacts/:contactid', db.contacts);
 * definitions:
 *   dataset:
 *     properties:
-*       Datasets:
-*         type: string
-*       
+*       type: object
 */
 
 /**
@@ -233,8 +120,7 @@ router.get('/v2/data/contacts/:contactid', db.contacts);
 */
 
 router.get('/v2/data/dataset/', db.dataset);
-router.get('/v2/data/dataset/:id', db.dataset);
-router.get('/v2/data/datasets/:datasetid/publications', db.publicationbydataset);
+router.get('/v2/data/dataset/:datasetid', db.dataset);
 
 /**
 * @swagger
@@ -270,62 +156,29 @@ router.get('/v2/data/datasets/:datasetid/publications', db.publicationbydataset)
 router.get('/v2/data/dbtables/', db.dbtables);
 router.get('/v2/data/dbtables/:table', db.dbtables);
 
-/**
-* @swagger
-* definitions:
-*   download:
-*     properties:
-*       type: object
-*/
-
-/**
-* @swagger
-* /v2/data/download:
-*   get:
-*     summary: Returns the named Neotoma Database table.
-*     description: Returns the named Neotoma Database table.
-*     parameters:
-*       - name: datasetid
-*         description: Table name.
-*         in: path
-*         required: true
-*         type: string
-*     produces:
-*       - application/json
-*     responses:
-*       200:
-*        description: Returned download object.
-*        schema:
-*          type: array
-*          items:
-*            $ref: '#/definitions/download'
-*/
-
-router.get('/v2/data/download/', db.download);
-router.get('/v2/data/download/:datasetid', db.download);
 
 /**
 * @swagger
 * definitions:
 *   geopolitical:
 *     properties:
-*       GeoPoliticalID:
+*       geopoliticalid:
 *         type: integer
 *         format: int32
-*       HigherGeoPoliticalID:
-*         type: integer
-*         format: int32
-*       Rank:
-*         type: integer
-*         format: int32
-*       GeoPoliticalUnit:
+*       geopoliticalname:
 *         type: string
-*       GeoPoliticalName:
+*       geopoliticalunit:
 *         type: string
-*       RecDateCreated: 
+*       rank:
+*         type: integer
+*         format: int32
+*       highergeopoliticalid:
+*         type: integer
+*         format: int32
+*       recdatecreated: 
 *         type: string
 *         format: dateTime
-*     	RecDateModified:
+*     	recdatemodified:
 *     	  type: string
 *     	  format: dateTime
 */
@@ -372,200 +225,8 @@ router.get('/v2/data/download/:datasetid', db.download);
 *             $ref: '#/definitions/geopolitical'
 */
 
-
 router.get('/v2/data/geopoliticalunits', db.geopoliticalunits);
-router.get('/v2/data/geopoliticalunits/:gpid', db.geopoliticalunits);
-router.get('/v2/data/sites/:siteid/geopoliticalunits', db.geopolbysite); // Geopolitical Units by site.
-
-/**
-* @swagger
-* definitions:
-*   occurrence:
-*     properties:
-*       OccurID: 
-*         type: integer
-*         format: int64
-*       Age:
-*         type: number
-*         format: float
-*       TaxonID:
-*         type: integer
-*         format: int32
-*       TaxonName:
-*         type: string
-*       AgeYounger:
-*         type: number
-*         format: float
-*       AgeOlder:
-*         type: number
-*         format: float
-*       DatasetID:
-*         type: integer
-*         format: int32
-*       DatabaseName:
-*         type: string
-*       DatasetType:
-*         type: string
-*       SiteName:
-*         type: string
-*       SiteID: 
-*         type: integer
-*         format: int32
-*       Altitude:
-*         type: number
-*         format: float
-*       LongitudeWest:
-*         type: number
-*         format: float
-*       LongitudeEast:
-*         type: number
-*         format: float
-*       LatitudeNorth:
-*         type: number
-*         format: float
-*       LatitudeSouth:
-*         type: number
-*         format: float
-*/
-
-
-/**
-* @swagger
-* /v2/data/occurrence:
-*   get:
-*     summary: Individual occurrence records for Neotoma records.
-*     description: Returns occurrence information for a particular taxon, geographic region or temporal slice.
-*     parameters:
-*       - name: occid
-*         description: Unique occurrence identifier.
-*         in: path
-*         required: true
-*         type: integer
-*         format: int64
-*     produces:
-*       - application/json
-*     responses:
-*       200:
-*        description: Occurrence
-*        schema:
-*          type: object
-*          items:
-*            $ref: '#/definitions/occurrence'
-*/
-
-router.get('/v2/data/occurrence/', db.occurrence);
-router.get('/v2/data/occurrence/:id', db.occurrence);
-
-/**
-* @swagger
-* definitions:
-*   pollen:
-*     properties:
-*       TaxonName:
-*         type: string
-*       EcolGroups:
-*         type: array
-*         items:
-*           type: string
-*       TaxonCode: 
-*         type: string
-*       Author:
-*         type: string
-*       PublicationID:
-*         type: integer
-*         format: int32
-*       TaxonID:
-*         type: integer
-*         format: int32
-*       TaxaGroupID:
-*         type: string
-*       HigherTaxonID:
-*         type: integer
-*         format: int32
-*       Extinct:
-*         type: boolean
-*       Notes:
-*         type: string
-*/
-
-/**
- * @swagger
- * /v2/data/pollen:
- *   get:
- *     summary: Pollen data from Neotoma.
- *     description: Returns information about pollen.
- *     parameters:
- *       - name: taxonid
- *         description: Numeric ID for taxa.
- *         in: path
- *         required: false
- *         type: integer
- *         format: int32
- *       - name: taxonname
- *         description: Name of the taxon, can include wildcards.
- *         in: query
- *         required: false
- *         type: string
- *       - name: nametype
- *         description: Not sure.
- *         in: query
- *         required: false
- *         type: string
- *       - name: taxonids
- *         description: Numeric ID for taxa.
- *         in: query
- *         required: false
- *         type: array
- *         items:
- *           type: integer
- *           format: int32
- *       - name: ageyoung
- *         description: Most recent age, in calendar years before present.
- *         in: query
- *         required: false
- *         type: number
- *         format: float
- *       - name: ageold
- *         description: Oldest age, in calendar years before present.
- *         in: query
- *         required: false
- *         type: number
- *         format: float
- *       - name: agedocontain
- *       - name: coords
- *       - name: bbox
- *       - name: wkt
- *       - name: sitename
- *         description: Name of site for pollen record.
- *         in: query
- *         required: false
- *         type: string
- *       - name: siteid
- *         description: Numeric site ID.
- *         in: query
- *         required: false
- *         type: integer
- *         format: int32
- *       - name: format
- *       - name: limit
- *       - name: offset
- *       - name: fields
- *       - name: vocab
- *       - name: occurrences
- *     produces:
- *       - application/json
- *     responses:
- *       200:
-*         description: A pollen response..
-*         schema:
-*           type: array
-*           items:
-*             $ref: '#/definitions/pollen'
-*/
-
-router.get('/v2/data/pollen/', db.pollen);
-router.get('/v2/data/pollen/:id', db.pollen);
-
+router.get('/v2/data/geopoliticalunits/:gpid', db.geopoliticalbyid);
 
 /**
 * @swagger
@@ -628,7 +289,7 @@ router.get('/v2/data/dataset/:datasetid/publications', db.publicationbydataset);
 /**
 * @swagger
 * definitions:
-*   sites:
+*   site:
 *     properties:
 *       siteid:
 *         type: integer
@@ -641,7 +302,7 @@ router.get('/v2/data/dataset/:datasetid/publications', db.publicationbydataset);
 *         type: string
 *       altitude:
 *         type: float
-*       CollectionUnits:
+*       collectionunits:
 *         type: object
 *         properties:
 *           collectionunit:
@@ -668,9 +329,9 @@ router.get('/v2/data/dataset/:datasetid/publications', db.publicationbydataset);
  * /v2/data/sites:
  *   get:
  *     summary: Site information.
- *     description: Returns information about Neotoma publications
+ *     description: Returns information about Neotoma sites
  *     parameters:
- *       - name: pubid
+ *       - name: siteid
  *         description: Numeric ID for publications.
  *         in: path
  *         required: false
@@ -680,7 +341,7 @@ router.get('/v2/data/dataset/:datasetid/publications', db.publicationbydataset);
  *         in: query
  *         required: false
  *         type: integer
- *       - name: siteid
+ *       - name: site
  *         description: Related site identifier.
  *         in: query
  *         required: false
@@ -698,7 +359,7 @@ router.get('/v2/data/dataset/:datasetid/publications', db.publicationbydataset);
 
 router.get('/v2/data/sites/', db.sitesquery); // Goes to the queries.
 router.get('/v2/data/sites/:siteid', db.sitesbyid); // Takes integers, including comma separated
-router.get('/v2/data/datasets/:datasetid/site', db.sitesbydataset); // Takes a dataset ID.
+router.get('/v2/data/dataset/:datasetid/site', db.sitesbydataset); // Takes a dataset ID.
 //router.get('/v2/data/publications/:pubid/site', db.sitesbypublication);
 router.get('/v2/data/geopoliticalunits/:gpid/site', db.sitesbygeopol);
 //router.get('/v2/data/contacts/:contactid/site', db.sitesbycontacts);
@@ -732,10 +393,10 @@ router.get('/v2/data/geopoliticalunits/:gpid/site', db.sitesbygeopol);
  * /v2/data/taxa:
  *   get:
  *     summary: Taxonomic information.
- *     description: Returns information about a taxon and (if requested) related taxa.
+ *     description: Returns information about a taxon and related taxa if requested.
  *     parameters:
  *       - name: taxonid
- *         description: Numeric ID for taxa.
+ *         description: Numeric ID for taxa, or comma separated string.
  *         in: path
  *         required: false
  *         type: integer
